trying to draw1

+++ lib/libvec/vec_init.c
@@ -6,13 +6,13 @@
 /*   By: scharuka <scharuka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/16 10:45:41 by scharuka          #+#    #+#             */
-/*   Updated: 2024/08/16 10:48:54 by scharuka         ###   ########.fr       */
+/*   Updated: 2024/08/16 12:03:17 by scharuka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector.h"
 
-t_vector	vec_init(float x, float y, float z)
+t_vector	vec_init(double x, double y, double z)
 {
 	t_vector	res;
